callbacks: Add message event callback.
Add support for the message event. The callback
does not receive all the parameters in the raw
JSON event, but the raw event is one of the parameters
if less commonly used info is needed.

A sample program is added demonstrating this callback.

Also fixes a bug where the client would disconnect
upon receiving an unknown event: we no longer check
the return value of slack_parse_message since we
should continue even if it fails.