slack-client.c: Fix reconnects and struct misalignment.
The callbacks struct was not always the same size,
due to one of its members only be defined conditionally.
As a result, this would change the offset into the struct
of successive members. A dummy member is now defined to
preserve memory offsets to ensure correct offsets are used
when dereferencing members.

Additionally, the reconnect logic in slack-client.c
was faulty for two reasons. First, the WebSocket reconnect
URL as sent in the request should contain only the URI
portion. Secondly, the wss_client was only created when
the Slack client was created, but it should be created
for each connection to ensure a new one is set up on reconnects.
Previously, this did not happen and so the wss_client would
be NULL, causing a segfault.