slack-client.c: Adjust autoreconnect timing.
Even with autoreconnect enabled, the example programs were
exiting if for some reason the WebSocket connection closed,
due to it being "too soon" if the last connection was attempted
in the last 300 seconds (5 minutes). A value of 15 seconds
works better in experimental testing, as permanent failures
such as invalidated cookies still cause it to abort, while
more innocuous disconnects can be worked around with a reconnect.